requirements.txt
streamlit
langchain
langchain_community
langchain-huggingface==0.1.2
langchain-ollama==0.2.3
langchain-openai==0.3.4
pypdf==5.3.0
python-dotenv
distro==1.9.0
jiter==0.8.2
joblib==1.4.2
mpmath==1.3.0
networkx==3.4.2
ollama==0.4.7
openai==1.61.1
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.1
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.8.0
torch==2.6.0
faiss-cpu